# [BUGFIX] PromQL: Noinline kahanSumInc, to avoid compiler optimisations