Observed on go1.24.4 darwin/arm64, the compensation variable is not correctly calculated.

Note you may have to run the tests several times to see it, to get the right ordering of series.

Since the problem went away when I tried to run under the debugger (which disables optimisations), and went away when I added sufficient printf statements (which discourages inlining), I hit upon the idea of disabling inlining for this one function.

See https://github.com/prometheus/prometheus/issues/16714#issuecomment-3092355237

It's going to slow things down a bit, but turns out `IsInf` is pretty slow already.

Fixes #16714.