@darshanime 2025-07-26
> It's going to slow things down a bit

should we measure how much slower? if the impact is too large, maybe we can disable the potential `inline` only for `*/arm64`?