@bboreham 2025-07-26
Doesn't show up in the benchmarks.  I added `sum_over_time(a_X[1d])`, where the inner loop calls Kahan.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Core(TM) i7-14700K
                                                                                                          │ 3602785a8.txt │           358499f86.txt            │
                                                                                                          │    sec/op     │    sec/op     vs base              │
RangeQuery/expr=a_one,steps=1-28                                                                              4.949µ ± 8%   5.331µ ±  1%       ~ (p=0.093 n=6)
RangeQuery/expr=a_one,steps=1000-28                                                                           28.88µ ± 1%   28.47µ ±  1%  -1.43% (p=0.002 n=6)
RangeQuery/expr=a_hundred,steps=1-28                                                                          139.5µ ± 1%   140.1µ ±  2%       ~ (p=0.818 n=6)
RangeQuery/expr=a_hundred,steps=1000-28                                                                       2.449m ± 1%   2.455m ±  0%       ~ (p=0.699 n=6)
RangeQuery/expr=rate(a_one[1m]),steps=1-28                                                                    7.819µ ± 1%   7.848µ ±  2%       ~ (p=0.082 n=6)
RangeQuery/expr=rate(a_one[1m]),steps=1000-28                                                                 71.14µ ± 0%   71.20µ ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-28                                                                168.4µ ± 1%   170.4µ ±  1%  +1.19% (p=0.015 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-28                                                             6.990m ± 1%   6.971m ±  1%       ~ (p=1.000 n=6)
RangeQuery/expr=rate(a_one[1m]),steps=10000-28                                                                711.3µ ± 2%   716.4µ ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-28                                                            76.10m ± 1%   76.14m ±  2%       ~ (p=0.699 n=6)
RangeQuery/expr=rate(sparse[1m]),steps=10000-28                                                               4.194m ± 0%   4.204m ±  2%       ~ (p=0.310 n=6)
RangeQuery/expr=double_exponential_smoothing(a_one[1d],_0.3,_0.3),steps=1-28                                  359.4µ ± 1%   359.9µ ±  2%       ~ (p=0.589 n=6)
RangeQuery/expr=double_exponential_smoothing(a_one[1d],_0.3,_0.3),steps=1000-28                               32.82m ± 0%   32.82m ±  0%       ~ (p=0.818 n=6)
RangeQuery/expr=double_exponential_smoothing(a_hundred[1d],_0.3,_0.3),steps=1-28                              25.47m ± 1%   25.51m ±  0%       ~ (p=0.937 n=6)
RangeQuery/expr=double_exponential_smoothing(a_hundred[1d],_0.3,_0.3),steps=1000-28                            3.245 ± 0%    3.246 ±  0%       ~ (p=0.394 n=6)
RangeQuery/expr=changes(a_one[1d]),steps=1-28                                                                 328.0µ ± 2%   329.1µ ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=changes(a_one[1d]),steps=1000-28                                                              40.74m ± 0%   40.76m ±  0%       ~ (p=1.000 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-28                                                             27.11m ± 1%   26.87m ±  1%  -0.89% (p=0.026 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-28                                                           3.577 ± 0%    3.575 ±  0%       ~ (p=0.485 n=6)
RangeQuery/expr=rate(a_one[1d]),steps=1-28                                                                    256.5µ ± 1%   256.3µ ±  1%       ~ (p=0.937 n=6)
RangeQuery/expr=rate(a_one[1d]),steps=1000-28                                                                 6.150m ± 0%   6.154m ±  1%       ~ (p=0.699 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-28                                                                20.14m ± 0%   20.08m ±  0%       ~ (p=0.132 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-28                                                             609.1m ± 1%   605.5m ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-28                                                        251.5µ ± 2%   251.9µ ±  2%       ~ (p=0.699 n=6)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-28                                                     3.589m ± 1%   3.579m ±  0%       ~ (p=0.394 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-28                                                    19.42m ± 1%   19.48m ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-28                                                 350.7m ± 0%   349.5m ±  0%  -0.34% (p=0.004 n=6)
RangeQuery/expr=-a_one,steps=1-28                                                                             6.267µ ± 1%   6.275µ ±  1%       ~ (p=1.000 n=6)
RangeQuery/expr=-a_one,steps=1000-28                                                                          30.37µ ± 2%   30.34µ ±  1%       ~ (p=0.485 n=6)
RangeQuery/expr=-a_hundred,steps=1-28                                                                         142.1µ ± 1%   142.2µ ±  1%       ~ (p=0.937 n=6)
RangeQuery/expr=-a_hundred,steps=1000-28                                                                      2.525m ± 1%   2.513m ±  0%       ~ (p=0.310 n=6)
RangeQuery/expr=a_one_-_b_one,steps=1-28                                                                      10.90µ ± 1%   10.88µ ±  1%       ~ (p=1.000 n=6)
RangeQuery/expr=a_one_-_b_one,steps=1000-28                                                                   227.9µ ± 3%   228.7µ ±  0%       ~ (p=0.310 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-28                                                              341.4µ ± 3%   346.6µ ±  1%       ~ (p=0.093 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-28                                                           16.23m ± 3%   16.62m ±  1%  +2.39% (p=0.041 n=6)
RangeQuery/expr=a_one_-_b_one,steps=10000-28                                                                  2.266m ± 1%   2.285m ±  1%  +0.84% (p=0.004 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-28                                                          173.0m ± 2%   175.3m ±  1%       ~ (p=0.065 n=6)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-28                                                     16.03µ ± 1%   16.08µ ±  1%       ~ (p=0.329 n=6)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-28                                                  89.66µ ± 1%   90.70µ ±  1%  +1.15% (p=0.026 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-28                                             245.7µ ± 2%   251.5µ ±  1%  +2.35% (p=0.002 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-28                                          8.708m ± 4%   8.838m ±  2%       ~ (p=0.310 n=6)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-28                                                      16.16µ ± 1%   16.25µ ±  1%       ~ (p=0.132 n=6)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-28                                                   111.4µ ± 2%   113.3µ ±  1%  +1.77% (p=0.015 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-28                                              260.1µ ± 2%   266.8µ ±  1%  +2.57% (p=0.004 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-28                                           10.86m ± 1%   10.82m ±  2%       ~ (p=0.937 n=6)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-28                                                  16.21µ ± 4%   16.23µ ±  0%       ~ (p=0.699 n=6)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-28                                               110.9µ ± 2%   111.8µ ±  1%       ~ (p=0.132 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-28                                          246.3µ ± 1%   248.9µ ±  1%  +1.02% (p=0.015 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-28                                       8.710m ± 4%   8.782m ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=a_one_and_b_one{l='notfound'},steps=1-28                                                      8.986µ ± 1%   9.034µ ±  2%       ~ (p=0.093 n=6)
RangeQuery/expr=a_one_and_b_one{l='notfound'},steps=1000-28                                                   82.55µ ± 1%   82.69µ ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=1-28                                              146.6µ ± 1%   145.5µ ±  2%       ~ (p=0.485 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=1000-28                                           3.306m ± 3%   3.199m ±  2%  -3.24% (p=0.009 n=6)
RangeQuery/expr=abs(a_one),steps=1-28                                                                         7.152µ ± 2%   7.100µ ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=abs(a_one),steps=1000-28                                                                      88.58µ ± 1%   88.84µ ±  1%       ~ (p=0.310 n=6)
RangeQuery/expr=abs(a_hundred),steps=1-28                                                                     162.5µ ± 1%   163.0µ ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=abs(a_hundred),steps=1000-28                                                                  6.005m ± 3%   6.019m ±  2%       ~ (p=0.937 n=6)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1-28                                      11.68µ ± 3%   11.45µ ±  3%  -1.92% (p=0.026 n=6)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1000-28                                   35.22µ ± 1%   35.20µ ±  0%       ~ (p=0.418 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1-28                                  167.1µ ± 2%   167.7µ ±  2%       ~ (p=0.310 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1000-28                               2.475m ± 1%   2.485m ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1-28                                             9.811µ ± 1%   9.833µ ±  1%       ~ (p=0.848 n=6)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1000-28                                          33.70µ ± 1%   33.66µ ±  1%       ~ (p=0.937 n=6)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1-28                                         156.7µ ± 1%   157.6µ ±  1%       ~ (p=0.485 n=6)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1000-28                                      2.460m ± 1%   2.464m ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=sum(a_one),steps=1-28                                                                         6.837µ ± 1%   6.803µ ±  0%       ~ (p=0.087 n=6)
RangeQuery/expr=sum(a_one),steps=1000-28                                                                      58.83µ ± 1%   58.86µ ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=sum(a_hundred),steps=1-28                                                                     136.1µ ± 2%   137.1µ ±  1%       ~ (p=0.180 n=6)
RangeQuery/expr=sum(a_hundred),steps=1000-28                                                                  2.818m ± 2%   2.840m ±  1%       ~ (p=0.065 n=6)
RangeQuery/expr=avg(a_one),steps=1-28                                                                         6.785µ ± 8%   6.792µ ±  2%       ~ (p=0.589 n=6)
RangeQuery/expr=avg(a_one),steps=1000-28                                                                      59.68µ ± 5%   59.25µ ±  0%       ~ (p=0.310 n=6)
RangeQuery/expr=avg(a_hundred),steps=1-28                                                                     137.4µ ± 0%   136.5µ ±  1%  -0.70% (p=0.004 n=6)
RangeQuery/expr=avg(a_hundred),steps=1000-28                                                                  2.907m ± 1%   2.880m ±  1%  -0.93% (p=0.041 n=6)
RangeQuery/expr=sum_without_(l)(h_one),steps=1-28                                                             23.62µ ± 1%   23.20µ ±  2%  -1.79% (p=0.026 n=6)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-28                                                          418.1µ ± 1%   412.3µ ±  1%  -1.39% (p=0.002 n=6)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1-28                                                         1.569m ± 1%   1.555m ±  1%  -0.88% (p=0.009 n=6)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-28                                                      37.23m ± 1%   36.18m ±  2%  -2.81% (p=0.002 n=6)
RangeQuery/expr=sum_without_(le)(h_one),steps=1-28                                                            21.02µ ± 1%   20.97µ ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-28                                                         334.0µ ± 0%   332.0µ ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1-28                                                        1.601m ± 1%   1.580m ±  1%  -1.29% (p=0.026 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-28                                                     39.75m ± 1%   39.68m ± 11%       ~ (p=1.000 n=6)
RangeQuery/expr=sum_by_(l)(h_one),steps=1-28                                                                  20.94µ ± 2%   21.10µ ±  1%       ~ (p=0.310 n=6)
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-28                                                               333.8µ ± 1%   334.8µ ±  1%       ~ (p=0.589 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1-28                                                              1.590m ± 1%   1.604m ±  1%  +0.89% (p=0.026 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-28                                                           39.44m ± 2%   39.95m ±  2%       ~ (p=0.065 n=6)
RangeQuery/expr=sum_by_(le)(h_one),steps=1-28                                                                 23.46µ ± 1%   23.48µ ±  1%       ~ (p=0.699 n=6)
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-28                                                              416.9µ ± 1%   418.0µ ±  1%       ~ (p=0.485 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1-28                                                             1.575m ± 1%   1.565m ±  1%       ~ (p=0.093 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-28                                                          37.43m ± 2%   37.14m ±  2%       ~ (p=0.065 n=6)
RangeQuery/expr=count_values('value',_h_one),steps=100-28                                                     803.3µ ± 2%   803.9µ ±  1%       ~ (p=0.937 n=6)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-28                                                 117.0m ± 3%   115.8m ±  3%       ~ (p=0.485 n=6)
RangeQuery/expr=topk(1,_a_one),steps=1-28                                                                     8.258µ ± 2%   8.285µ ±  0%       ~ (p=0.394 n=6)
RangeQuery/expr=topk(1,_a_one),steps=1000-28                                                                  162.7µ ± 2%   163.6µ ±  2%       ~ (p=0.240 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=1-28                                                                 142.0µ ± 2%   141.9µ ±  1%       ~ (p=0.818 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=1000-28                                                              3.331m ± 0%   3.356m ±  1%       ~ (p=0.240 n=6)
RangeQuery/expr=topk(5,_a_one),steps=1-28                                                                     8.316µ ± 2%   8.320µ ±  1%       ~ (p=1.000 n=6)
RangeQuery/expr=topk(5,_a_one),steps=1000-28                                                                  162.6µ ± 5%   163.9µ ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=topk(5,_a_hundred),steps=1-28                                                                 144.7µ ± 3%   145.4µ ±  1%       ~ (p=0.310 n=6)
RangeQuery/expr=topk(5,_a_hundred),steps=1000-28                                                              4.341m ± 1%   4.376m ±  1%  +0.80% (p=0.026 n=6)
RangeQuery/expr=limitk(1,_a_one),steps=1-28                                                                   8.201µ ± 1%   8.229µ ±  0%       ~ (p=0.589 n=6)
RangeQuery/expr=limitk(1,_a_one),steps=1000-28                                                                159.4µ ± 2%   160.7µ ±  1%       ~ (p=0.394 n=6)
RangeQuery/expr=limitk(1,_a_hundred),steps=1-28                                                               139.7µ ± 1%   140.8µ ±  1%       ~ (p=0.065 n=6)
RangeQuery/expr=limitk(1,_a_hundred),steps=1000-28                                                            2.721m ± 1%   2.715m ±  2%       ~ (p=0.589 n=6)
RangeQuery/expr=limitk(5,_a_one),steps=1-28                                                                   8.172µ ± 1%   8.231µ ±  1%       ~ (p=0.145 n=6)
RangeQuery/expr=limitk(5,_a_one),steps=1000-28                                                                158.7µ ± 0%   160.8µ ±  1%  +1.35% (p=0.002 n=6)
RangeQuery/expr=limitk(5,_a_hundred),steps=1-28                                                               140.7µ ± 2%   141.4µ ±  1%       ~ (p=0.065 n=6)
RangeQuery/expr=limitk(5,_a_hundred),steps=1000-28                                                            2.852m ± 3%   2.877m ±  1%       ~ (p=0.132 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_one),steps=1-28                                                            8.144µ ± 5%   8.192µ ±  1%       ~ (p=0.132 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_one),steps=1000-28                                                         137.8µ ± 1%   138.7µ ±  1%       ~ (p=0.093 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_hundred),steps=1-28                                                        145.8µ ± 1%   147.0µ ±  2%  +0.79% (p=0.041 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_hundred),steps=1000-28                                                     4.849m ± 3%   4.900m ±  2%       ~ (p=0.065 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_one),steps=1-28                                                            8.098µ ± 2%   8.165µ ±  2%  +0.83% (p=0.041 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_one),steps=1000-28                                                         136.5µ ± 0%   138.7µ ±  1%  +1.62% (p=0.002 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_hundred),steps=1-28                                                        156.2µ ± 1%   157.8µ ±  2%  +1.04% (p=0.041 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_hundred),steps=1000-28                                                     6.570m ± 3%   6.711m ±  1%  +2.15% (p=0.041 n=6)
RangeQuery/expr=limit_ratio(-0.5,_a_one),steps=1-28                                                           8.381µ ± 1%   8.445µ ±  1%  +0.77% (p=0.041 n=6)
RangeQuery/expr=limit_ratio(-0.5,_a_one),steps=1000-28                                                        175.9µ ± 1%   177.8µ ±  1%  +1.10% (p=0.002 n=6)
RangeQuery/expr=limit_ratio(-0.5,_a_hundred),steps=1-28                                                       158.3µ ± 2%   160.1µ ±  1%       ~ (p=0.132 n=6)
RangeQuery/expr=limit_ratio(-0.5,_a_hundred),steps=1000-28                                                    6.817m ± 3%   6.982m ±  1%       ~ (p=0.093 n=6)
RangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=1-28                                                  15.20µ ± 4%   15.30µ ±  3%       ~ (p=0.937 n=6)
RangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=1000-28                                               300.6µ ± 1%   302.0µ ±  1%       ~ (p=0.180 n=6)
…